pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cad-preproc"
version = "0.1.0"
description = "Pipeline de pré-processamento CAD (DXF/DWG) para grafo topológico"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Seu Nome", email = "voce@example.com" }]
keywords = ["CAD", "DXF", "DWG", "graph", "topology", "AI", "architecture"]
dependencies = [
  "ezdxf>=1.2.0",
  "shapely>=2.0.0",
  "networkx>=3.0",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
  "black>=24.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "types-PyYAML",
]

[project.scripts]
cad-preproc = "cli:main"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_ignores = true
